#120: (Two RqFake instances are equal if they contain the same head and byte array body): Extracted the inner class for requests with byte array bodies to a separate file.
